发布代码生成、更新20+表单组件，优化数据字典，gf版本更新到2.3.1

+++ server/internal/websocket/model.go
@@ -18,7 +18,7 @@ type WRequest struct {
 type WResponse struct {
 	Event     string      `json:"event"`              // 事件名称
 	Data      interface{} `json:"data,omitempty"`     // 数据
-	Code      int64       `json:"code"`               // 状态码
+	Code      int         `json:"code"`               // 状态码
 	ErrorMsg  string      `json:"errorMsg,omitempty"` // 错误消息
 	Timestamp int64       `json:"timestamp"`          // 服务器时间
 }
